tests: Use a function to extract transaction fee
and minimize code duplication
